chore: move pagination helper to gitlab package
